reupload

+++ main.cpp
@@ -18,28 +18,28 @@ using namespace std;
 /**
  * calculates the sum of all positive integers between int and 1
  *
- * @param int num the starting number
- * @return int the sum of positive integers
+ * @param unsigned long int num the starting number
+ * @return unsigned long int the sum of positive integers
  * 
  */
-int sumIncremental(int num);
+unsigned long int sumIncremental(unsigned long int num);
 
 
 /**
  * calculates the product of all positive integers between int and 1
  *
- * @param int num the starting number
- * @return int the product of positive integers
+ * @param unsigned long int num the starting number
+ * @return unsigned long int the product of positive integers
  * 
  */
-int productIncremental(int num);
+unsigned long int productIncremental(unsigned long int num);
 
 
 int main() {
 
   cout << "Enter a positive integer: ";
 
-  int num = 0;
+  unsigned long int num = 0;
   cin >> num;
   cout << "The sum of positive integers between 1 and " << num << " is " << sumIncremental(num) << "." << endl;
   cout << "The product of these integers is " << productIncremental(num) << "." << endl;
@@ -48,21 +48,23 @@ int main() {
 }
 
 
-int sumIncremental(int num) {
-  int sum = 0;
+unsigned long int sumIncremental(unsigned long int num) {
+  unsigned long int sum = 0;
 
-  //add num to total, then decrease num by 1
+  // add num to total, then decrease num by 1
   while (num > 0) sum += num--;
   
   return sum;
+
 }
 
 
-int productIncremental(int num) {
-  int product = 1;
+unsigned long int productIncremental(unsigned long int num) {
+  unsigned long int product = 1;
 
-  //multiply product by num, then decrease num by 1
+  // multiply product by num, then decrease num by 1
   while (num > 1) product *= num--;
 
   return product;
+
 }
